Get & report the in & out degree of sampled nodes in the original graph
In and out degree of a node in the sampled graph and the original graph are different, because the sampled graph may not include all the neighbors of the node in the original graph. However, the in and out degree of node in the original graph are significant representative features, so including them along with the in and out degree in the sampled graph.